requirements.txt
fastapi
uvicorn
fastapi-cache2==0.2.1